requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
environs==9.5.0
Faker==12.0.1
geographiclib==2.0
geopy==2.4.0
gunicorn==21.2.0
hypothesis==6.92.1
importlib-metadata==7.0.1
iniconfig==2.0.0
itypes==1.2.0
marshmallow==3.20.1
mixer==7.2.2
packaging==23.2
Pillow==10.1.0
pipwin==0.5.2
pluggy==1.3.0
psycopg==3.1.17
psycopg-binary==3.1.17
pycparser==2.21
pyjsparser==2.7.1
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
setproctitle==1.3.3
sortedcontainers==2.4.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
whitenoise==6.6.0
zipp==3.17.0
